Fix(plugins): Prevent deprecation warnings when deprecated filters are not used
