+++ .pre-commit-config.yaml
@@ -180,9 +180,10 @@ repos:
         name: Build documentation for collection filter plugins.
         entry: ansible-doc-extractor --template ansible_collections/arista/avd/docs/templates/plugin-docs.j2 --markdown "ansible_collections/arista/avd/docs/plugins/Filter_plugins/"
         language: python
-        types: [python]
+        types_or: [python, yaml]
         additional_dependencies: ['ansible-doc-extractor>=0.1.10', 'ansible-core>=2.15.0,<2.18.0']
         files: ansible_collections/arista/avd/plugins/filter/
+        exclude: ^(ansible_collections/arista/avd/plugins/filter/deprecated_filters.py)$
 
       - id: docs-plugin-lookup
         name: Build documentation for collection lookup plugins.
