+++ ansible_collections/arista/avd/plugins/filter/generate_esi.yml
@@ -1,0 +1,31 @@
+---
+DOCUMENTATION:
+  name: generate_esi
+  collection: arista.avd
+  author: Arista Ansible Team (@aristanetworks)
+  version_added: "1.1"
+  short_description: Transforms short_esi `0303:0202:0101` to EVPN ESI format `0000:0000:0303:0202:0101`
+  description: Concatenates the given `esi_prefix` and `short_esi`.
+  positional: _input
+  options:
+    _input:
+      description: Short ESI value as per AVD definition in eos_designs.
+      type: string
+      required: true
+    esi_prefix:
+      description: ESI prefix value. Will be concatenated with the `short_esi`.
+      type: string
+      default: "0000:0000:"
+  deprecated:
+    removed_in: "5.0.0"
+    why: This filter is no longer used by AVD and is very simple to replace with generic Jinja syntax.
+    alternative: Use Jinja string concatenation instead like `{{ <esi_prefix> ~ <short_esi> }}`
+
+EXAMPLES: |
+  ---
+  esi: "{{ short_esi | arista.avd.generate_esi('deaf:beed:') }}"
+
+RETURN:
+  _value:
+    description: Concatenated string of `esi_prefix` and `short_esi` like `0000:0000:0303:0202:0101`
+    type: string
