+++ ansible_collections/arista/avd/plugins/filter/generate_lacp_id.yml
@@ -1,0 +1,27 @@
+---
+DOCUMENTATION:
+  name: generate_lacp_id
+  collection: arista.avd
+  author: Arista Ansible Team (@aristanetworks)
+  version_added: "1.1"
+  short_description: Transforms short_esi `0303:0202:0101` to LACP ID format `0303.0202.0101`
+  description: Replaces `:` with `.`
+  positional: _input
+  options:
+    _input:
+      description: Short ESI value as per AVD definition in eos_designs.
+      type: string
+      required: true
+  deprecated:
+    removed_in: "5.0.0"
+    why: This filter is no longer used by AVD and is very simple to replace with a generic Jinja filter.
+    alternative: Use the builtin `replace` filter instead like `{{ <short_esi> | replace(':', '.') }}`
+
+EXAMPLES: |-
+  ---
+  lacp_id: "{{ short_esi | arista.avd.generate_lacp_id }}"
+
+RETURN:
+  _value:
+    description: String based on LACP ID format like 0303.0202.0101
+    type: string
