+++ ansible_collections/arista/avd/plugins/filter/generate_route_target.yml
@@ -1,0 +1,29 @@
+---
+DOCUMENTATION:
+  name: generate_route_target
+  collection: arista.avd
+  author: Arista Ansible Team (@aristanetworks)
+  version_added: "1.1"
+  short_description: Transforms short_esi `0303:0202:0101` to route-target format `03:03:02:02:01:01`
+  description: Removes `:` and inserts new `:` for each two characters.
+  positional: _input
+  options:
+    _input:
+      description: Short ESI value as per AVD definition in eos_designs.
+      type: string
+      required: true
+  deprecated:
+    removed_in: "5.0.0"
+    why: This filter is no longer used by AVD and is very simple to replace with a generic Jinja filter.
+    alternative: |-
+      Use the builtin `ansible.builtin.regex_replace` filter instead like
+      `{{ <short_esi> | ansible.builtin.regex_replace('(\\d{2})(\\d{2}):(\\d{2})(\\d{2}):(\\d{2})(\\d{2})', '\\1:\\2:\\3:\\4:\\5:\\6') }}`
+
+EXAMPLES: |-
+  ---
+  rt: "{{ short_esi | arista.avd.generate_route_target }}"
+
+RETURN:
+  _value:
+    description: String based on route-target format like 03:03:02:02:01:01
+    type: string
